+++ src/main/java/absolutelyaya/ultracraft/mixin/LivingEntityMixin.java
@@ -489,7 +489,7 @@ public abstract class LivingEntityMixin extends Entity implements LivingEntityAc
 			return;
 		if(!getWorld().isClient)
 			ExplosionHandler.explosion(player, player.getWorld(), player.getPos(), DamageSources.get(player.getWorld(),
-					DamageSources.KNUCKLE_BLAST, player), 1f, 0.75f, 6, false);
+					DamageSources.KNUCKLE_BLAST, player), 1f, 0.75f, 6, true);
 		else
 		{
 			if(player instanceof WingedPlayerEntity winged && winged.isSliding())